Optimize parts page for better performance
Replace img tag with Next.js Image component and extract mock search data to reduce bundle size.

- Replace img with Next.js Image for automatic optimization
- Extract mock parts search logic to separate file
- Reduce inline code in parts page component
- Target: reduce bundle size from 37.4 kB to under 25 kB
